move fri and polyops to files
