+++ crates/prover/src/core/backend/gpu/mod.rs
@@ -3,6 +3,8 @@ pub mod m31;
 mod bit_reverse;
 mod column;
 mod accumulation;
+mod circle;
+mod fri;
 
 use std::fmt::Debug;
 use std::sync::Arc;
@@ -13,17 +15,12 @@ use cudarc::nvrtc::compile_ptx;
 use once_cell::sync::Lazy;
 
 use self::m31::LoadPackedBaseField;
-use super::{Backend, Col};
-use crate::core::circle::{CirclePoint, Coset};
+use super::{Backend};
 use crate::core::fields::m31::{BaseField, P};
 use crate::core::fields::qm31::SecureField;
-use crate::core::fri::FriOps;
 use crate::core::pcs::quotients::{ColumnSampleBatch, QuotientOps};
-use crate::core::poly::circle::{
-    CanonicCoset, CircleDomain, CircleEvaluation, CirclePoly, PolyOps, SecureEvaluation,
+use crate::core::poly::circle::{CircleDomain, CircleEvaluation, SecureEvaluation,
 };
-use crate::core::poly::line::LineEvaluation;
-use crate::core::poly::twiddles::TwiddleTree;
 use crate::core::poly::BitReversedOrder;
 const VECTOR_SIZE: usize = 16;
 
@@ -55,66 +52,6 @@ struct GpuBackend;
 impl Backend for GpuBackend {}
 
 
-impl PolyOps for GpuBackend {
-    type Twiddles = Vec<BaseField>;
-
-    fn new_canonical_ordered(
-        _coset: CanonicCoset,
-        _values: Col<Self, BaseField>,
-    ) -> CircleEvaluation<Self, BaseField, BitReversedOrder> {
-        todo!()
-    }
-
-    fn interpolate(
-        _eval: CircleEvaluation<Self, BaseField, BitReversedOrder>,
-        _itwiddles: &TwiddleTree<Self>,
-    ) -> CirclePoly<Self> {
-        todo!()
-    }
-
-    fn eval_at_point(_poly: &CirclePoly<Self>, _point: CirclePoint<SecureField>) -> SecureField {
-        todo!()
-    }
-
-    fn extend(_poly: &CirclePoly<Self>, _log_size: u32) -> CirclePoly<Self> {
-        todo!()
-    }
-
-    fn evaluate(
-        _poly: &CirclePoly<Self>,
-        _domain: CircleDomain,
-        _twiddles: &TwiddleTree<Self>,
-    ) -> CircleEvaluation<Self, BaseField, BitReversedOrder> {
-        todo!()
-    }
-
-    fn precompute_twiddles(_coset: Coset) -> TwiddleTree<Self> {
-        todo!()
-    }
-}
-
-impl FriOps for GpuBackend {
-    fn fold_line(
-        _eval: &LineEvaluation<Self>,
-        _alpha: SecureField,
-        _twiddles: &TwiddleTree<Self>,
-    ) -> LineEvaluation<Self> {
-        todo!()
-    }
-
-    fn fold_circle_into_line(
-        _dst: &mut LineEvaluation<GpuBackend>,
-        _src: &SecureEvaluation<Self>,
-        _alpha: SecureField,
-        _twiddles: &TwiddleTree<Self>,
-    ) {
-        todo!()
-    }
-
-    fn decompose(_eval: &SecureEvaluation<Self>) -> (SecureEvaluation<Self>, SecureField) {
-        todo!()
-    }
-}
 
 impl QuotientOps for GpuBackend {
     fn accumulate_quotients(
